requirements.txt
Flask
Pillow
flask-cors
